#44 TimerDM class created
